feat(frontend): add toast hook and component
